fix: resolve clippy warnings for CI
- Replace row.get(0) with row.first() across all files
- Use vec![] macro instead of Vec::new() + push
- Remove unnecessary u8 cast
- Fix manual_clamp, clone_on_copy, useless_conversion
- Use is_some_and instead of map_or(false, ...)
- Simplify iterator loops in read_only.rs
- Add allow attributes for too_many_arguments
- Derive Default instead of manual impl
- Allow uninlined_format_args project-wide

+++ src/cli/args.rs
@@ -826,7 +826,7 @@ fn parse_matches(matches: &ArgMatches) -> CliArgs {
         markdown: matches.get_flag("markdown"),
         pretty: matches.get_flag("pretty"),
     };
-    let verbose = matches.get_count("verbose") as u8;
+    let verbose = matches.get_count("verbose");
     let quiet = matches.get_flag("quiet");
 
     let command = match matches.subcommand() {
